fix: evaluate
