+++ package.json
@@ -18,7 +18,7 @@
     "next-auth": "^4.24.7",
     "next-mdx-remote": "^4.4.1",
     "next-pwa": "^5.6.0",
-    "openai": "^4.49.1",
+    "openai": "^4.52.4",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-to-print": "^2.15.1",
