+++ src/components/navbar/navbar.tsx
@@ -8,7 +8,7 @@ import user from "@/../public/icons/user.svg";
 import iep from "@/../public/icons/iep.svg";
 import manual from "@/../public/icons/manual.svg";
 import add from "@/../public/icons/chat-add.svg";
-import assesment from "@/../public/icons/assesment.svg"
+import evaluate from "@/../public/icons/evaluate.svg"
 import UserInfo from "./info/user/user";
 import ClassInfo from "./info/class/class";
 import Menu from "@/components/navbar/menu/menu";
@@ -76,11 +76,11 @@ export default function Navbar({
               name="manual"
             />
             <Menu
-              link={"/class/" + classid + "/assesment"}
-              src={assesment}
-              alt="assesment icon"
+              link={"/class/" + classid + "/evaluate"}
+              src={evaluate}
+              alt="evaluate icon"
               title="성취도 평가"
-              name="assesment"
+              name="evaluate"
             />
           </div>
           <div>
